+++ src/components/trip-days.js
@@ -1,0 +1,7 @@
+const tripDaysTemplate = () => {
+  return (
+    `<ul class="trip-days"></ul>`
+  );
+};
+
+export {tripDaysTemplate};
