+++ src/components/trip-info.js
@@ -1,0 +1,7 @@
+const tripInfoTemplate = () => {
+  return (
+    `<section class="trip-main__trip-info  trip-info"></section>`
+  );
+};
+
+export {tripInfoTemplate};
